+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `ROUND(number, [digits])` formula function (#119)
+
 ## [0.5.1](https://github.com/garritfra/cell/compare/v0.5.0...v0.5.1) - 2026-06-29
 
 ### Added